invoice_scanner_monetizable_fixed/requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
flask-wtf==1.1.1
wtforms==3.0.1
